@@ -1,114 +1,53 @@
 {
-  "name": "template",
+  "name": "lbdserver",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@babel/core": "^7.16.0",
-    "@comunica/actor-init-sparql": "^1.22.3",
-    "@date-io/dayjs": "^2.13.1",
-    "@emotion/react": "^11.7.1",
-    "@emotion/styled": "^11.6.0",
-    "@esbuild-plugins/node-globals-polyfill": "^0.1.1",
-    "@esbuild-plugins/node-modules-polyfill": "^0.1.4",
-    "@i2labs/dns": "^1.0.1",
-    "@inrupt/solid-client-authn-browser": "^1.11.3",
-    "@inrupt/vocab-common-rdf": "^1.0.3",
-    "@mui/icons-material": "^5.3.1",
-    "@mui/lab": "^5.0.0-alpha.73",
-    "@mui/material": "^5.4.0",
-    "@mui/x-data-grid": "^5.5.1",
-    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.3",
-    "@react-three/drei": "^8.20.0",
-    "@react-three/fiber": "^6.0.13",
-    "@solid/lit-prism-patterns": "^0.13.9",
-    "@svgr/webpack": "^5.5.0",
-    "@xeokit/xeokit-sdk": "^1.9.0",
-    "browserify-fs": "^1.0.0",
-    "browserify-zlib": "^0.2.0",
-    "browserslist": "^4.18.1",
-    "buffer": "^6.0.3",
-    "case-sensitive-paths-webpack-plugin": "^2.4.0",
-    "crypto-browserify": "^3.12.0",
-    "css-loader": "^6.5.1",
-    "css-minimizer-webpack-plugin": "^3.2.0",
+    "@emotion/react": "^11.9.0",
+    "@emotion/styled": "^11.8.1",
+    "@mui/lab": "^5.0.0-alpha.77",
+    "@mui/material": "^5.6.1",
+    "@react-three/drei": "^9.4.2",
+    "@react-three/fiber": "^8.0.11",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.0.1",
+    "@testing-library/user-event": "^13.5.0",
+    "@xeokit/xeokit-sdk": "^2.2.5-beta-4",
     "dayjs": "^1.11.0",
-    "https-browserify": "^1.0.0",
-    "identity-obj-proxy": "^3.0.0",
-    "jsonld-streaming-serializer": "^1.3.0",
-    "lbdserver-client-api": "^0.0.21",
-    "lodash": "^4.17.21",
-    "mem": "^9.0.2",
-    "microdata-rdf-streaming-parser": "^1.2.0",
-    "mini-css-extract-plugin": "^2.4.5",
-    "net-browserify": "^0.2.4",
-    "os-browserify": "^0.3.0",
-    "path-browserify": "^1.0.1",
-    "process": "^0.11.10",
-    "querystring-es3": "^0.2.1",
-    "rdf-data-factory": "^1.1.0",
-    "rdf-isomorphic": "^1.3.0",
-    "rdf-literal": "^1.3.0",
-    "rdf-store-stream": "^1.3.0",
-    "rdf-string": "^1.6.0",
-    "rdf-string-ttl": "^1.2.0",
-    "rdf-terms": "^1.7.1",
-    "rdfa-streaming-parser": "^1.5.0",
-    "rdfxml-streaming-parser": "^1.5.0",
-    "react": "^17.0.2",
-    "react-dayjs": "^0.3.2",
-    "react-dev-utils": "^12.0.0",
-    "react-dom": "^17.0.2",
-    "react-draggable": "^4.4.4",
-    "react-grid-layout": "^1.3.3",
-    "react-refresh": "^0.11.0",
-    "react-router-dom": "^6.2.1",
-    "readable-web-to-node-stream": "^3.0.2",
-    "recoil": "^0.6.1",
-    "relative-to-absolute-iri": "^1.0.6",
-    "resolve": "^1.20.0",
-    "resolve-url-loader": "^4.0.0",
-    "semver": "^7.3.5",
-    "single-spa": "^5.9.3",
-    "source-map-loader": "^3.0.1",
-    "sparqlalgebrajs": "^4.0.2",
-    "sparqljson-parse": "^1.7.0",
-    "sparqljson-to-tree": "^2.1.0",
-    "sparqlxml-parse": "^1.5.0",
-    "stream-browserify": "^3.0.0",
-    "stream-http": "^3.2.0",
-    "style-loader": "^3.3.1",
+    "express": "^4.17.3",
+    "npm-run-all": "^4.1.5",
+    "prop-types": "^15.8.1",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-scripts": "5.0.1",
     "suncalc": "^1.9.0",
-    "tls-browserify": "^0.2.2",
-    "url": "^0.11.0",
-    "util": "^0.12.4"
-  },
-  "browser": {
-    "child_process": false,
-    "dns": false,
-    "http2": false
-  },
-  "devDependencies": {
-    "@babel/core": "^7.14.5",
-    "@babel/plugin-transform-runtime": "^7.12.1",
-    "@babel/preset-env": "^7.14.5",
-    "@babel/preset-react": "^7.14.5",
-    "babel-loader": "^8.2.2",
-    "clean-webpack-plugin": "^3.0.0",
-    "graphql-ld-comunica-solid": "^1.0.1",
-    "html-webpack-plugin": "^4.5.0",
-    "terser-webpack-plugin": "^5.2.5",
-    "ts-loader": "^9.2.6",
-    "typescript": "^4.5.5",
-    "web-vitals": "^2.1.4",
-    "webpack": "^5.38.1",
-    "webpack-cli": "^4.9.2",
-    "webpack-dev-server": "^3.11.2",
-    "webpack-manifest-plugin": "^4.0.2",
-    "webpack-merge": "^5.2.0",
-    "workbox-webpack-plugin": "^6.4.1"
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
-    "start": "webpack serve --config config/webpack.dev.js",
-    "build": "webpack --config config/webpack.prod.js"
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject",
+    "api": "node server.js",
+    "app": "npm start",
+    "all": "npm-run-all --parallel api app"
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
